backend/foodgram/requirements.txt
Django==2.2.28
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==4.0.1
gunicorn==20.0.4
psycopg2-binary==2.8.6